requirements.txt
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0